samplers and some traits
